Moving MockInteractor to the client
