+++ maestro-cli/src/main/java/maestro/cli/command/mockserver/MockServerProjectIdCommand.kt
@@ -2,7 +2,7 @@ package maestro.cli.command.mockserver
 
 import maestro.cli.util.PrintUtils
 import maestro.cli.util.PrintUtils.err
-import maestro.studio.MockInteractor
+import maestro.mockserver.MockInteractor
 import picocli.CommandLine.Command
 import picocli.CommandLine.ParentCommand
 import java.util.concurrent.Callable
